api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
requests==2.31.0
openai==1.10.0
xmltodict==0.13.0
markdown==3.5.2
reportlab==4.0.9
jinja2==3.1.3
aiofiles==23.2.1
httpx==0.26.0

